+++ litmus-portal/frontend/.eslintignore
@@ -2,3 +2,4 @@
 **/*.md
 **/cypress/
 **/geo.json
+**/*.spec.tsx
