+++ litmus-portal/frontend/src/components/ProgressBar/LinearProgressBar/LinearProgressBar.spec.tsx
@@ -1,7 +1,7 @@
 /// <reference types="Cypress" />
 import { mount } from 'cypress-react-unit-test';
 import React from 'react';
-import LinearProgressBar from '../../src/components/ProgressBar/LinearProgressBar';
+import LinearProgressBar from '../LinearProgressBar';
 
 describe('Linear Progressbar Testing', () => {
   it('Progressbar stroke for value 3', () => {
